requirements.txt
pika
requests
pyyaml